third_party/WebKit/LayoutTests/fast/deprecated-flexbox/child-flexing-expected.txt

Check if box-flex specified flexboxes expand or shrink correctly. If you see any scrollbars, then the test has failed.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

vertically expanding
PASS element.scrollHeight is 100
PASS element.scrollWidth is 100
vertically shrinking
PASS element.scrollHeight is 100
PASS element.scrollWidth is 100
horizontally expanding
PASS element.scrollHeight is 100
PASS element.scrollWidth is 100
horizontally shrinking
PASS element.scrollHeight is 100
PASS element.scrollWidth is 100
PASS successfullyParsed is true

TEST COMPLETE
